refactor: remove unnecessery comments
